refactor :: 피그백 반영 #24
