--- src/App.tsx
@@ -1,7 +1,0 @@
-import HomePage from "./pages";
-
-function App() {
-  return;
-}
-
-export default App;
