+++ src/Root.tsx
@@ -1,7 +1,7 @@
 import { QueryClientProvider, QueryClient } from "react-query";
 import { RecoilRoot } from "recoil";
-import PageTemplagte from "./components/common/PageTemplate/PageTemplate";
-import ThemeProviderContainer from "./components/common/ThemeProviderContainer/ThemeProviderContainer";
+import PageTemplagte from "./components/Common/PageTemplate/PageTemplate";
+import ThemeProviderContainer from "./components/Common/ThemeProviderContainer/ThemeProviderContainer";
 import HomePage from "./pages";
 
 const queryClient = new QueryClient();
