+++ src/lib/axios/customAxios.ts
@@ -3,8 +3,8 @@ import config from "../../config/config.json";
 import {
   ACCESS_TOKEN_KEY,
   REQUEST_TOKEN_KEY,
-} from "../../constants/token/token.constant";
-import token from "../token/token";
+} from "../../constants/Token/token.constant";
+import token from "../Token/token";
 import { customAxiosInterceptor } from "./interceptor";
 
 export const dodamV6Axios = axios.create({
