+++ src/lib/token/token.ts
@@ -1,9 +1,9 @@
-import cookie from "../cookie/cookie";
+import cookie from "../Cookie/cookie";
 import jwt, { JwtPayload } from "jsonwebtoken";
 import {
   ACCESS_TOKEN_KEY,
   REFRESH_TOKEN_KEY,
-} from "../../constants/token/token.constant";
+} from "../../constants/Token/token.constant";
 
 class Token {
   public getToken(key: string): string | undefined {
