+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import * as S from "./style";
-import Schedule from "src/components/home/ScheduleCalendar";
-import ScheduleList from "src/components/home/ScheduleList";
+import Schedule from "src/components/Home/ScheduleCalendar";
+import ScheduleList from "src/components/Home/ScheduleList";
 
 const HomePage = () => {
   return (
