+++ src/queries/member/member.query.ts
@@ -1,7 +1,7 @@
 import { AxiosError } from "axios";
 import { useQuery, UseQueryResult } from "react-query";
-import memberRepository from "../../repository/member/member.repository";
-import { MyMemberResponse } from "../../types/member/member.type";
+import memberRepository from "../../repository/Member/member.repository";
+import { MyMemberResponse } from "../../types/Member/member.type";
 
 export const useGetMember = (): UseQueryResult<MyMemberResponse, AxiosError> =>
   useQuery("member/getMyMember", () => memberRepository.getMyMember(), {
