+++ src/repository/member/member.repository.ts
@@ -1,5 +1,5 @@
-import { dodamV6Axios } from "../../lib/axios/customAxios";
-import { MyMemberResponse } from "../../types/member/member.type";
+import { dodamV6Axios } from "../../lib/Axios/customAxios";
+import { MyMemberResponse } from "../../types/Member/member.type";
 
 class MemberRepository {
   public async getMyMember(): Promise<MyMemberResponse> {
@@ -8,4 +8,4 @@ class MemberRepository {
   }
 }
 const memberRepository = new MemberRepository();
-export default  memberRepository;
+export default memberRepository;
