+++ src/types/token/token.type.ts
@@ -1,4 +1,4 @@
-import { Response } from "../util/response.type";
+import { Response } from "../Util/response.type";
 
 export interface RefreshResponse extends Response {
   data: string;
